refactor: improve file structure
