+++ public/index.html
@@ -6,15 +6,15 @@
         content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>myLibrary</title>
-  <link rel="stylesheet" href="./styles.css">
-  <link rel="icon" href="./media/icons/logo.png">
+  <link rel="stylesheet" href="/styles.css">
+  <link rel="icon" href="/media/icons/logo.png">
 </head>
 <body>
 
 <div id="container">
   <header id="home-header">
     <div class="home-header-center left-side">
-      <img class="logo" src="media/icons/logo.svg" alt="logo">
+      <img class="logo" src="/media/icons/logo.svg" alt="logo">
       <h1>myLibrary</h1>
     </div>
   </header>
@@ -61,5 +61,5 @@
 </div>
 
 </body>
-<script src="./homepage-script.js"></script>
+<script src="/js/homepage-script.js"></script>
 </html>
